moving metric UI to sepearate module
